+++ src/cargo_home.ts
@@ -2,4 +2,10 @@ import * as path from 'path';
 import * as os from 'os';
 import workspaceConfig from './workspace_config';
 
-export const cargoHome = workspaceConfig().cargoHome || process.env.CARGO_HOME || path.resolve(os.homedir(), '.cargo');
+export function cargoHome () {
+return workspaceConfig().cargoHome || process.env.CARGO_HOME || path.resolve(os.homedir(), '.cargo');
+} 
+
+export function isCargoNextestEnabled() {
+return workspaceConfig().nextest.enable;
+}