+++ src/is_nextest_installed.ts
@@ -1,8 +1,11 @@
 import * as fs from 'fs';
 import * as path from 'path';
-import { cargoHome } from './cargo_home';
+import { cargoHome, isCargoNextestEnabled } from './cargo_home';
 
 export default async function isCargoNextestInstalled(): Promise<boolean> {
-	const cargoNextestPath = path.join(cargoHome, 'bin', 'cargo-nextest');
+	if (!isCargoNextestEnabled()) {
+		return false;
+	}
+	const cargoNextestPath = path.join(cargoHome(), 'bin', 'cargo-nextest');
 	return fs.existsSync(cargoNextestPath);
 }